refactor: 깃허브 클라이언트 패키지 변경
